+++ command-queue-sample/src/main/java/com/zhuinden/commandqueuesample/MainViewModel.kt
@@ -1,9 +1,9 @@
 package com.zhuinden.commandqueuesample
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
 import android.os.Handler
 import android.os.Looper
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.zhuinden.commandqueue.CommandQueue
 import java.util.concurrent.Executors